backend/requirement.txt
functions-framework==3.*
gspread
pandas
scikit-learn
google-auth